Positive and Negative sentence prediction using bert/requirement.txt.txt
pytorch
flask
numpy
pandas
transformers
